backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
nltk==3.8.1
openai==1.30.1
pinecone-client==3.1.0
pinecone-client[grpc]==3.1.0
pip==23.2.1
python-dotenv==1.0.1
PyYAML==6.0.1
scipy==1.12.0
tiktoken==0.6.0
gunicorn==21.2.0
pysbd==0.3.4
rq==1.16.1
marshmallow==3.21.1
loguru==0.7.2
google-generativeai==0.6.0